Move structs and functions related to the PGS format into new separate file
